fix(linux): attach GTK drag handler to all windows, not just main
New windows opened via open_new_window() were not draggable because
the GTK button-press-event handler was only attached to the 'main'
window in setup. Extract handler into drag::attach_drag_handler() and
call it for every new WebviewWindow on Linux.
